exec() chef-client to resume the run after updating the omnibus install
Rather than kill the chef client and leave it to the user to pick up the
pieces and, in most cases, simply resume the run, do that by default by
exec'ing chef-client again in the ruby_block to start using the new
version.

The original behavior is still accessible via attributes.